move all data in a separate folder
