+++ src/data/projects.fr.ts
